requirements.txt
paho-mqtt==1.4.0
wyzesense==0.0.4
